workaround for newlines in PDFList
Added code to work around a problem with newline characters in the results of the PDFList script. Some lines in the "linedata" input variable would have an incorrect ending, which made phantomjs.exe hang and eventually time out without any sort of error message. This change should make the "ZIP: P + M + Links" downloads work much better, at least giving us some results back instead of an empty zip. A better fix would be to find the source of the newlines in the PDFList script - if we do that, we can remove this workaround.