converted from UTF-16
BETAExtract.cpp had some annoying encoding, making Git treat it as a binary file and generally making it harder to work with. Saved it out with regular 8 bit ANSI encoding without changing anything. Hopefully no strings got messed up by doing this, I did check for that by converting it back to UTF-16 locally and confirming that it was identical to the original file, so at least no information is lost. I think it's best to convert this now, otherwise it will just be harder to see the changes we make.